+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.6] - 2025-11-14
+
 ### Added
 
 - Added Portrait for `Portrait_Artemis_Default_01`
@@ -63,7 +65,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added Wiggle effects for Aphrodite
 - Added Glit effects for Hera
 
-[unreleased]: https://github.com/Kuipo/Hades-2-Free-the-Nipple-Edition/compare/1.1.5...HEAD
+[unreleased]: https://github.com/Kuipo/Hades-2-Free-the-Nipple-Edition/compare/1.1.6...HEAD
+[1.1.6]: https://github.com/Kuipo/Hades-2-Free-the-Nipple-Edition/compare/1.1.5...1.1.6
 [1.1.5]: https://github.com/Kuipo/Hades-2-Free-the-Nipple-Edition/compare/1.1.4...1.1.5
 [1.1.4]: https://github.com/Kuipo/Hades-2-Free-the-Nipple-Edition/compare/1.1.3...1.1.4
 [1.1.3]: https://github.com/Kuipo/Hades-2-Free-the-Nipple-Edition/compare/1.1.2...1.1.3